Move Collection models to wagtail.core.models.collections
